refactor(profile): decouple profile identity from account identity
BREAKING CHANGE: Profile now uses its own ProfileId instead of AccountId as Primary Key.
This enables 1:N relationship between Accounts and Profiles.

Changes:
- Domain: Introduced ProfileId ValueObject and updated Profile aggregate.
- Domain: Renamed 'username' to 'handle' to align with social domain language.
- Persistence: Migrated PostgreSQL schema to use (id, region_code) as Composite PK.
- Persistence: Unified Identity, Stats, and Cache via UnifiedProfileRepository.
- Kernel: Added Slug ValueObject for shared validation logic (Handle/Username).
- Infrastructure: Updated Kafka consumer to handle account.created events with the new identity mapping.
- API: Updated gRPC mappers and handlers to reflect handle/id changes.feat(profile): add post_count to profile stats with ScyllaDB integration

- Update ProfileStats domain value object to include post_count.
- Add post_count field to profile.v1.ProfileStats protobuf message.
- Update ScyllaProfileRepository to fetch and save post_count using counters.
- Update gRPC mappers to handle the new field.
- Refactor integration tests structure for better stability with Redis.

Architecture decision: post_count is stored in ScyllaDB alongside other
social counters to ensure write scalability and consistent data locality
within the CompositeProfileRepository.

+++ crates/account/src/domain/builders/account_builder.rs
@@ -17,7 +17,7 @@ pub struct AccountBuilder {
     locale: Option<Locale>,
     phone: Option<PhoneNumber>,
     birth_date: Option<BirthDate>,
-    version: i32,
+    version: u64,
     last_active_at: Option<DateTime<Utc>>,
 }
 
@@ -59,7 +59,7 @@ impl AccountBuilder {
         account_state: AccountState,
         birth_date: Option<BirthDate>,
         locale: Locale,
-        version: i32,
+        version: u64,
         created_at: DateTime<Utc>,
         updated_at: DateTime<Utc>,
         last_active_at: Option<DateTime<Utc>>,
